refactor gulp project definitions
