+++ ui/chat/gulpfile.js
@@ -1,51 +1,3 @@
-const gulp = require('gulp');
-const source = require('vinyl-source-stream');
-const buffer = require('vinyl-buffer');
-const colors = require('ansi-colors');
-const logger = require('fancy-log');
-const watchify = require('watchify');
-const browserify = require('browserify');
-const uglify = require('gulp-uglify');
-const size = require('gulp-size');
-const tsify = require('tsify');
+const lilaGulp = require('../gulp/tsProject.js');
 
-const browserifyOpts = (debug) => ({
-  entries: ['./src/main.ts'],
-  standalone: 'LichessChat',
-  debug: debug
-});
-const baseFileName = "lichess.chat";
-const destination = () => gulp.dest('../../public/compiled/');
-
-const prod = () => browserify(browserifyOpts(false))
-    .plugin(tsify)
-    .bundle()
-    .pipe(source(`${baseFileName}.min.js`))
-    .pipe(buffer())
-    .pipe(uglify())
-    .pipe(size())
-    .pipe(destination());
-
-const dev = () => browserify(browserifyOpts(true))
-    .plugin(tsify)
-    .bundle()
-    .pipe(source(`${baseFileName}.js`))
-    .pipe(destination());
-
-const watch = () => {
-
-  const bundle = () => bundler.bundle()
-    .on('error', error => logger.error(colors.red(error.message)))
-    .pipe(source(`${baseFileName}.js`))
-    .pipe(destination());
-
-  const bundler = watchify(
-    browserify(browserifyOpts(true)).plugin(tsify)
-  ).on('update', bundle).on('log', logger.info);
-
-  return bundle();
-};
-
-gulp.task('prod', prod);
-gulp.task('dev', dev);
-gulp.task('default', watch);
+lilaGulp('LichessChat', 'lichess.chat', __dirname);
