+++ ui/simul/gulpfile.js
@@ -1,45 +1,3 @@
-const gulp = require('gulp');
-const source = require('vinyl-source-stream');
-const buffer = require('vinyl-buffer');
-const colors = require('ansi-colors');
-const logger = require('fancy-log');
-const watchify = require('watchify');
-const browserify = require('browserify');
-const uglify = require('gulp-uglify');
-const size = require('gulp-size');
+const lilaGulp = require('../gulp/jsProject.js');
 
-const browserifyOpts = (debug) => ({
-  entries: ['./src/main.js'],
-  standalone: 'LichessSimul',
-  debug: debug
-});
-const destination = () => gulp.dest('../../public/compiled/');
-
-const prod = () => browserify(browserifyOpts(false)).bundle()
-  .pipe(source('lichess.simul.min.js'))
-  .pipe(buffer())
-  .pipe(uglify())
-  .pipe(size())
-  .pipe(destination());
-
-const dev = () => browserify(browserifyOpts(true)).bundle()
-  .pipe(source('lichess.simul.js'))
-  .pipe(destination());
-
-const watch = () => {
-
-  const bundle = () => bundler.bundle()
-    .on('error', error => logger.error(colors.red(error.message)))
-    .pipe(source('lichess.simul.js'))
-    .pipe(destination());
-
-  const bundler = watchify(
-    browserify(Object.assign({}, watchify.args, browserifyOpts(true)))
-  ).on('update', bundle).on('log', logger.info);
-
-  return bundle();
-};
-
-gulp.task('prod', prod);
-gulp.task('dev', dev);
-gulp.task('default', watch);
+lilaGulp('LichessSimul', 'lichess.simul', __dirname);
