+++ ui/simul/package.json
@@ -26,6 +26,7 @@
     "fancy-log": "^1",
     "uglify-js": "^3",
     "vinyl-source-stream": "^2",
+    "vinyl-buffer": "^1",
     "watchify": "^3"
   },
   "dependencies": {
